test(e2e): address review on the migrate smoke test
- Run the hops under pnpm rather than npm, relaxing peer strictness with
  pnpm's own `strict-peer-dependencies=false` instead of `legacy-peer-deps`.
- Scaffold with `internal#test-matrix` so each hop gets the start version's own
  full generator coverage. Releases predating that generator keep the fixed
  recipe; the fallback goes when the supported range no longer reaches them.
- Migrate from the last 5 releases rather than only the latest.
- Move `publishForMigrateSmokeTest` into `migrate-versions`, alongside the
  version resolution it depends on, and skip it unless the run includes the
  migrate test — every lane shares global-setup, and mirroring a release per
  start version is wasted work for the lanes that never read it.
- Drop the user-customisation assertion, which hardcoded a path from the
  recipe's website project and kept the driver from staying generic.

Assert the upgraded version against the installed package rather than the
manifest range: on a pnpm workspace the manifest holds `catalog:` and the
version lives in `pnpm-workspace.yaml`, so node_modules is the
package-manager-agnostic record — and what the migrations actually run from.
